# src/main/java/hexlet/code/app/dto/TaskCreateDTO.java
package hexlet.code.app.dto;

import com.fasterxml.jackson.annotation.JsonProperty;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Size;
import lombok.Getter;
import lombok.Setter;

import java.util.List;

@Getter
@Setter
public class TaskCreateDTO {

    private Long index;

    @Size(min = 1)
    private String title;

    private String content;

    @NotNull
    private String status;

    @JsonProperty("assignee_id")
    private Long assigneeId;
    //private Long assignee_id;

    private List<Long> taskLabelIds;
}