# src/main/java/hexlet/code/app/dto/TaskParamsDTO.java
package hexlet.code.app.dto;

import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
public class TaskParamsDTO {

    private String titleCont;

    private Long assigneeId;

    private String status;

    private Long labelId;
}