# src/main/resources/api-docs.yaml
openapi: 3.1.0
info:
  title: Spring OpenAPI example
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- BearerAuth: []
paths:
...
  /api/tasks/{id}:
    get:
      tags:
      - task-controller
      operationId: showTask
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TaskDTO"
    put:
      tags:
      - task-controller
      operationId: updateTask
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdateDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TaskDTO"
    delete:
      tags:
      - task-controller
      operationId: deleteTask
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: No Content
...
  /api/tasks:
    get:
      tags:
      - task-controller
      operationId: index
      parameters:
      - name: params
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/TaskParamsDTO"
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskDTO"
    post:
      tags:
      - task-controller
      operationId: createTask
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCreateDTO"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TaskDTO"
...
components:
  schemas:
    JsonNullableString:
      type: object
      properties:
        present:
          type: boolean
...
    JsonNullableListLong:
      type: object
      properties:
        present:
          type: boolean
    JsonNullableLong:
      type: object
      properties:
        present:
          type: boolean
    TaskUpdateDTO:
      type: object
      properties:
        index:
          $ref: "#/components/schemas/JsonNullableLong"
        title:
          $ref: "#/components/schemas/JsonNullableString"
        content:
          $ref: "#/components/schemas/JsonNullableString"
        status:
          $ref: "#/components/schemas/JsonNullableString"
        taskLabelIds:
          $ref: "#/components/schemas/JsonNullableListLong"
        assignee_id:
          $ref: "#/components/schemas/JsonNullableLong"
      required:
      - status
    TaskDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        index:
          type: integer
          format: int64
        title:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date
        status:
          type: string
        taskLabelIds:
          type: array
          items:
            type: integer
            format: int64
        assignee_id:
          type: integer
          format: int64
...
    TaskCreateDTO:
      type: object
      properties:
        index:
          type: integer
          format: int64
        title:
          type: string
          maxLength: 2147483647
          minLength: 1
        content:
          type: string
        status:
          type: string
        taskLabelIds:
          type: array
          items:
            type: integer
            format: int64
        assignee_id:
          type: integer
          format: int64
      required:
      - status
...
    TaskParamsDTO:
      type: object
      properties:
        titleCont:
          type: string
        assigneeId:
          type: integer
          format: int64
        status:
          type: string
        labelId:
          type: integer
          format: int64
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT